[feat/#26] Navigate

+++ app/src/main/java/com/sopt/korailtalk/data/remote/model/response/SeatTicketResponse.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class SeatTicketResponse(
-    @SerialName("ticketId") val ticketId: Int
+    @SerialName("ticketId") val ticketId: Long
 )
