+++ app/src/main/java/com/sopt/korailtalk/presentation/ui/seatmap/SeatSelectingState.kt
@@ -3,6 +3,6 @@ package com.sopt.korailtalk.presentation.ui.seatmap
 sealed class SeatSelectingState {
     data object Idle: SeatSelectingState()
     data object Loading: SeatSelectingState()
-    data class Success(val data: Int): SeatSelectingState()
+    data class Success(val data: Long): SeatSelectingState()
     data class Failure(val message: String): SeatSelectingState()
 }
